+++ src/main/java/insane96mcp/survivalreimagined/mixin/LivingEntityMixin.java
@@ -1,0 +1,37 @@
+package insane96mcp.survivalreimagined.mixin;
+
+import insane96mcp.insanelib.base.Feature;
+import insane96mcp.survivalreimagined.module.movement.feature.TerrainSlowdown;
+import net.minecraft.util.Mth;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.ai.attributes.Attributes;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Redirect;
+
+@Mixin(LivingEntity.class)
+public class LivingEntityMixin {
+
+    @Redirect(at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/LivingEntity;isSprinting()Z"), method = "jumpFromGround")
+    public boolean onSprintJumpCheck(LivingEntity instance) {
+        if (instance.isSprinting() && Feature.isEnabled(TerrainSlowdown.class)) {
+            float yRot = instance.getYRot() * ((float)Math.PI / 180F);
+            float boost = 0.2f;
+            boost *= getMovementSpeedRatio(instance);
+            instance.setDeltaMovement(instance.getDeltaMovement().add((-Mth.sin(yRot) * boost), 0.0D, (Mth.cos(yRot) * boost)));
+            return false;
+        }
+        return instance.isSprinting();
+    }
+
+    //TODO Move to Lib
+    private static double getMovementSpeedRatio(LivingEntity livingEntity) {
+        double baseMS = 0.1;
+        if (livingEntity.isSprinting()) {
+            baseMS += 0.029999999329447746;
+        }
+
+        double playerMS = livingEntity.getAttributeValue(Attributes.MOVEMENT_SPEED);
+        return playerMS / baseMS;
+    }
+}